When creditCard payment amount is equal to the total order amount the order is finished.
Added a parameter on method confirm of payment.
This parameter will be a flag to know when we can skip the confirmation
dialog when finalizing a order.

Added a new field on creditCard template object, the original
amount(amount remaining on order) to be compared with the actual credit
card paid amount.