New sacred files for the "best-sellers" page
Created new files for the sacred form of the best-sellers page. This
page is special and don't use the same structure of any other sacred
page/style.

All product images here (10 at all) are using a css image map  that
needs to be tested on tablets (iPad and android). Also the
cache/refresh of the "sacred-best-sellers.css" file needs to be tested.