All calendars have 3 visualisations modes (month/year with days; year with months; and years with years).
To see and change among this modes, just click on the second-top-bottom
(with the month/year).

--
Problem: when we change the visualization mode, also the box width is
changed. This way, for the second and third mode, we have some width
problems.

Now current and old version of calendars are with a min. width to
correct them when the visualization mode is changed.